backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
httpx==0.27.0
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
numpy==2.1.3
